feat(core): floating image action panel in inspector
Show a small floating toolbar below the inspector selection box when an
image is selected, with quick-access Replace and Crop icons. Extract
AssetPickerDialog into its own module and add an openReplace callback
to the inspector context so the floating panel and the side panel share
the same flow.
