+++ .changeset/inspector-image-quick-actions.md
@@ -1,0 +1,5 @@
+---
+"@open-slide/core": minor
+---
+
+Show a floating action panel below the inspector selection box when an image is selected, with quick-access Replace and Crop icons.
